refactor(logbook): extract LogbookVisitRow into own file
Move LogbookVisitRow and ConfidenceBadge out of LogbookView.swift into
LogbookVisitRow.swift so TripDetailView (Task 9) can reuse them directly.
